fix(dashboard): alias propagation in Maily components
