+++ apps/dashboard/src/components/variable/components/variable-icon.tsx
@@ -3,7 +3,7 @@ import { RiErrorWarningLine } from 'react-icons/ri';
 import { Code2 } from '@/components/icons/code-2';
 import { DigestVariableIcon } from '@/components/icons/digest-variable-icon';
 import { RepeatVariable } from '@/components/icons/repeat-variable';
-import { REPEAT_BLOCK_ITERABLE_ALIAS } from '@/components/workflow-editor/steps/email/variables/variables';
+import { REPEAT_BLOCK_ITERABLE_ALIAS } from '@/components/workflow-editor/steps/email/variables/repeat-block-aliases';
 import { DIGEST_PREVIEW_MAP } from '@/components/variable/utils/digest-variables';
 
 export const VariableIcon = ({
