+++ apps/dashboard/src/components/workflow-editor/steps/email/views/variable-view.tsx
@@ -7,7 +7,8 @@ import { IsAllowedVariable, LiquidVariable } from '@/utils/parseStepVariables';
 import { NodeViewProps } from '@tiptap/core';
 import { NodeViewWrapper } from '@tiptap/react';
 import { useCallback, useMemo, useState } from 'react';
-import { resolveRepeatBlockAlias } from '../variables/variables';
+import { VariableFrom } from '../variables/variables';
+import { resolveRepeatBlockAlias } from '../variables/repeat-block-aliases';
 import { DIGEST_VARIABLES_ENUM, getDynamicDigestVariable } from '@/components/variable/utils/digest-variables';
 import { useWorkflow } from '@/components/workflow-editor/workflow-provider';
 import { useWorkflowSchema } from '@/components/workflow-editor/workflow-schema-provider';
